update the mainwindows
