--- src/gui/MainWindows.cpp
@@ -1,89 +1,0 @@
-#include "MainWindow.h"
-#include <iostream>
-
-static void glfw_error_callback(int error, const char* description) {
-    fprintf(stderr, "Glfw Error %d: %s\n", error, description);
-}
-
-MainWindow::MainWindow(int width, int height, const char* title) 
-    : clear_color(ImVec4(0.45f, 0.55f, 0.60f, 1.00f)) {
-
-    glfwSetErrorCallback(glfw_error_callback);
-    if (!glfwInit())
-        throw std::runtime_error("Failed to initialize GLFW");
-
-#if defined(__APPLE__)
-    const char* glsl_version = "#version 150";
-    glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 3);
-    glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 2);
-    glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);
-    glfwWindowHint(GLFW_OPENGL_FORWARD_COMPAT, GL_TRUE);
-#else
-    const char* glsl_version = "#version 130";
-    glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 3);
-    glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 0);
-#endif
-
-    window = glfwCreateWindow(width, height, title, NULL, NULL);
-    if (!window)
-        throw std::runtime_error("Failed to create GLFW window");
-
-    glfwMakeContextCurrent(window);
-    glfwSwapInterval(1);  // Enable vsync
-
-    InitializeImGui();
-}
-
-MainWindow::~MainWindow() {
-    Cleanup();
-    glfwDestroyWindow(window);
-    glfwTerminate();
-}
-
-void MainWindow::InitializeImGui() {
-    IMGUI_CHECKVERSION();
-    ImGui::CreateContext();
-    // ImPlot::CreateContext();
-    // ImGui::StyleColorsDark();
-    ImGui_ImplGlfw_InitForOpenGL(window, true);
-    ImGui_ImplOpenGL3_Init("#version 130");
-}
-
-void MainWindow::Render() {
-    // New frame for ImGui
-    ImGui_ImplOpenGL3_NewFrame();
-    ImGui_ImplGlfw_NewFrame();
-    ImGui::NewFrame();
-
-    // Show ImGui windows and elements here
-    ImGui::Begin("Hello, world!");
-    ImGui::Text("This is some useful text.");
-    ImGui::End();
-
-    // Rendering
-    ImGui::Render();
-    int display_w, display_h;
-    glfwGetFramebufferSize(window, &display_w, &display_h);
-    glViewport(0, 0, display_w, display_h);
-    glClearColor(clear_color.x * clear_color.w,
-                 clear_color.y * clear_color.w,
-                 clear_color.z * clear_color.w,
-                 clear_color.w);
-    glClear(GL_COLOR_BUFFER_BIT);
-    ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
-    glfwSwapBuffers(window);
-}
-
-void MainWindow::RunMainLoop() {
-    while (!glfwWindowShouldClose(window)) {
-        glfwPollEvents();
-        Render();
-    }
-}
-
-void MainWindow::Cleanup() {
-    ImGui_ImplOpenGL3_Shutdown();
-    ImGui_ImplGlfw_Shutdown();
-    // ImPlot::DestroyContext();
-    // ImGui::DestroyContext();
-}
