+++ src/gui/main_windows.cpp
@@ -1,0 +1,79 @@
+#include "main_window.h"
+#include <stdexcept>
+#include <stdio.h>
+
+static void glfw_error_callback(int error, const char *description) {
+  fprintf(stderr, "Glfw Error %d: %s\n", error, description);
+}
+
+MainWindow::MainWindow(int width, int height, const char *title) {
+
+  glfwSetErrorCallback(glfw_error_callback);
+  if (!glfwInit())
+    // return 1;
+    throw std::runtime_error("Failed to initialize GLFW");
+
+// Decide GL+GLSL versions
+#if defined(__APPLE__)
+  const char *glsl_version = "#version 150";
+  glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 3);
+  glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 2);
+  glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROILE);
+  glfwWindowHint(GLFW_OPENGL_FORWARD_COMPAT, GL_TRUE);
+#else
+  const char *glsl_version = "#version 130";
+  glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 3);
+  glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 0);
+#endif
+
+  window = glfwCreateWindow(width, height, title, NULL, NULL);
+  if (!window)
+    throw std::runtime_error("Failed to create GLFW window");
+
+  glfwMakeContextCurrent(window);
+  glfwSwapInterval(1); // Enable vsync
+
+  IMGUI_CHECKVERSION();
+  ImGui::CreateContext();
+  // ImPlot::CreateContext();
+  // ImGui::StyleColorsDark();
+
+  // Setup Dear ImGui style
+  ImGui::StyleColorsDark();
+  // ImGui::StyleColorsLight();
+
+  // Setup Platform/Renderer backends
+  ImGui_ImplGlfw_InitForOpenGL(window, true);
+  ImGui_ImplOpenGL3_Init(glsl_version);
+};
+
+MainWindow::~MainWindow() {
+  ImGui_ImplOpenGL3_Shutdown();
+  ImGui_ImplGlfw_Shutdown();
+
+  glfwDestroyWindow(window);
+  glfwTerminate();
+}
+
+void MainWindow::runMainLoop() {
+  while (!glfwWindowShouldClose(window)) {
+    glfwPollEvents();
+
+    ImGui_ImplOpenGL3_NewFrame();
+    ImGui_ImplGlfw_NewFrame();
+    ImGui::NewFrame();
+
+    ImGui::Render();
+    int display_w, display_h;
+    glfwGetFramebufferSize(window, &display_w, &display_h);
+    glViewport(0, 0, display_w, display_h);
+    // glClearColor(clear_color.x * clear_color.w,
+    //              clear_color.y * clear_color.w,
+    //              clear_color.z * clear_color.w,
+    //              clear_color.w);
+    //
+    glClear(GL_COLOR_BUFFER_BIT);
+    ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
+    glfwSwapBuffers(window);
+  }
+}
